Move API tests to src/tests location
