--- src/app/space/wizard/services/forge-wizard.api.mock.ts
@@ -1,246 +1,0 @@
-export const step_1_1_response =  {
-  'metadata': {
-    'deprecated': false,
-    'category': 'Uncategorized',
-    'name': 'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-    'description': 'No Description'
-  },
-  'state': {
-    'valid': true,
-    'canExecute': false,
-    'wizard': true,
-    'canMoveToNextStep': true,
-    'canMoveToPreviousStep': false,
-    'steps': [
-      'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-      'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'
-    ]
-  },
-  'inputs': [
-    {
-      'name': 'gitOrganisation',
-      'shortName': ' ',
-      'valueType': 'io.fabric8.forge.generator.git.GitOrganisationDTO',
-      'inputType': 'org.jboss.forge.inputType.DEFAULT',
-      'enabled': true,
-      'required': true,
-      'deprecated': false,
-      'label': 'Organisation',
-      'description': 'The github organisation to import repositories from',
-      'valueChoices': [
-        {
-          'id': 'corinnekrych',
-          'description': 'My personal github account',
-          'htmlUrl': 'https://github.com/corinnekrych',
-          'name': 'corinnekrych',
-          'valid': true
-        }
-      ],
-      'class': 'UISelectOne',
-      'value': 'corinnekrych'
-    }
-  ]
-};
-
-export const step_1_2_input = {
-  'metadata': {
-    'deprecated': false,
-    'category': 'Uncategorized',
-    'name': 'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-    'description': 'No Description'
-  },
-  'state': {
-    'valid': true,
-    'canExecute': false,
-    'wizard': true,
-    'canMoveToNextStep': true,
-    'canMoveToPreviousStep': false,
-    'steps': [
-      'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-      'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'
-    ]
-  },
-  'inputs': [
-    {
-      'name': 'gitOrganisation',
-      'shortName': ' ',
-      'valueType': 'io.fabric8.forge.generator.git.GitOrganisationDTO',
-      'inputType': 'org.jboss.forge.inputType.DEFAULT',
-      'enabled': true,
-      'required': true,
-      'deprecated': false,
-      'label': 'Organisation',
-      'description': 'The github organisation to import repositories from',
-      'valueChoices': [
-        {
-          'id': '3musket33rs',
-          'avatarUrl': 'https://avatars2.githubusercontent.com/u/1974314?v=4',
-          'description': 'http://3musket33rs.github.com',
-          'htmlUrl': 'https://github.com/3musket33rs',
-          'name': '3.musket33rs',
-          'valid': true
-        },
-        {
-          'id': 'corinnekrych',
-          'description': 'My personal github account',
-          'htmlUrl': 'https://github.com/corinnekrych',
-          'name': 'corinnekrych',
-          'valid': true
-        }
-      ],
-      'class': 'UISelectOne',
-      'value': 'corinnekrych'
-    }
-  ],
-  'stepIndex': 0
-};
-
-export const step_1_2_response = {
-  'state': {
-    'valid': true,
-    'canExecute': false,
-    'wizard': true,
-    'canMoveToNextStep': true,
-    'canMoveToPreviousStep': false,
-    'steps': [
-      'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-      'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'
-    ]
-  },
-  'messages': [],
-  'inputs': [
-    {
-      'name': 'gitOrganisation',
-      'shortName': ' ',
-      'valueType': 'io.fabric8.forge.generator.git.GitOrganisationDTO',
-      'inputType': 'org.jboss.forge.inputType.DEFAULT',
-      'enabled': true,
-      'required': true,
-      'deprecated': false,
-      'label': 'Organisation',
-      'description': 'The github organisation to import repositories from',
-      'valueChoices': [
-        {
-          'id': '3musket33rs',
-          'avatarUrl': 'https://avatars2.githubusercontent.com/u/1974314?v=4',
-          'description': 'http://3musket33rs.github.com',
-          'htmlUrl': 'https://github.com/3musket33rs',
-          'name': '3.musket33rs',
-          'valid': true
-        },
-        {
-          'id': 'corinnekrych',
-          'description': 'My personal github account',
-          'htmlUrl': 'https://github.com/corinnekrych',
-          'name': 'corinnekrych',
-          'valid': true
-        }
-      ],
-      'class': 'UISelectOne',
-      'value': 'corinnekrych'
-    }
-  ]
-};
-
-export const step_1_3_input = {
-  'metadata': {
-    'deprecated': false,
-    'category': 'Uncategorized',
-    'name': 'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-    'description': 'No Description'
-  },
-  'state': {
-    'valid': true,
-    'canExecute': false,
-    'wizard': true,
-    'canMoveToNextStep': true,
-    'canMoveToPreviousStep': false,
-    'steps': [
-      'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-      'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'
-    ]
-  },
-  'inputs': [
-    {
-      'name': 'gitOrganisation',
-      'shortName': ' ',
-      'valueType': 'io.fabric8.forge.generator.git.GitOrganisationDTO',
-      'inputType': 'org.jboss.forge.inputType.DEFAULT',
-      'enabled': true,
-      'required': true,
-      'deprecated': false,
-      'label': 'Organisation',
-      'description': 'The github organisation to import repositories from',
-      'valueChoices': [
-        {
-          'id': '3musket33rs',
-          'avatarUrl': 'https://avatars2.githubusercontent.com/u/1974314?v=4',
-          'description': 'http://3musket33rs.github.com',
-          'htmlUrl': 'https://github.com/3musket33rs',
-          'name': '3.musket33rs',
-          'valid': true
-        },
-        {
-          'id': 'corinnekrych',
-          'description': 'My personal github account',
-          'htmlUrl': 'https://github.com/corinnekrych',
-          'name': 'corinnekrych',
-          'valid': true
-        }
-      ],
-      'class': 'UISelectOne',
-      'value': '$selected_github_organisation'
-    }
-  ],
-  'stepIndex': 1
-};
-
-export const step_1_3_response = {
-  'metadata': {
-    'deprecated': false,
-    'category': 'Uncategorized',
-    'name': 'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep',
-    'description': 'No Description'
-  },
-  'state': {
-    'valid': false,
-    'canExecute': false,
-    'wizard': true,
-    'canMoveToNextStep': false,
-    'canMoveToPreviousStep': true,
-    'steps': [
-      'io.fabric8.forge.generator.github.GithubImportPickOrganisationStep',
-      'io.fabric8.forge.generator.github.GithubImportPickRepositoriesStep'
-    ]
-  },
-  'inputs': [
-    {
-      'name': 'gitRepositoryPattern',
-      'shortName': ' ',
-      'valueType': 'io.fabric8.forge.generator.git.GitRepositoryDTO',
-      'inputType': 'org.jboss.forge.inputType.DEFAULT',
-      'enabled': true,
-      'required': true,
-      'deprecated': false,
-      'label': 'Repository name pattern',
-      'description': 'The regex pattern to match repository names',
-      'valueChoices': [
-        {
-          'id': '3musket33rs.github.com',
-          'description': 'Web Site',
-          'name': '3musket33rs.github.com',
-          'valid': true
-        },
-        {
-          'id': 'corinnekrych',
-          'description': 'My personal github account',
-          'htmlUrl': 'https://github.com/corinnekrych',
-          'name': 'corinnekrych',
-          'valid': true
-        }
-      ],
-      'class': 'UISelectMany',
-      'value': []
-    }
-  ]
-};
