+++ src/tests/forge-api/forge-wizard.api.spec.ts
@@ -1,17 +1,15 @@
 import {async, TestBed} from '@angular/core/testing';
 const Pact = require('pact-web');
-
-import { step_1_1_response,
-  step_1_2_input,
-  step_1_2_response,
-  step_1_3_input,
-  step_1_3_response
-} from './forge-wizard.api.mock';
-import {HttpModule} from "@angular/http";
-import {Fabric8ForgeService} from "./fabric8-forge.service";
-import {LoggerFactory} from "../common/logger";
-import {AuthenticationService} from "ngx-login-client";
-import {ApiLocatorService} from "../../../shared/api-locator.service";
+import {HttpModule} from '@angular/http';
+import {LoggerFactory} from '../../app/space/wizard/common/logger';
+import {AuthenticationService} from 'ngx-login-client';
+import {ApiLocatorService} from '../../app/shared/api-locator.service';
+import {Fabric8ForgeService} from '../../app/space/wizard/services/fabric8-forge.service';
+import {step_1_1_output} from './import-wizard/step_1_1_output';
+import {step_1_3_input} from './import-wizard/step_1_3_input';
+import {step_1_3_output} from 'tests/forge-api/import-wizard/step_1_3_output';
+import {step_1_2_input} from './import-wizard/step_1_2_input';
+import {step_1_2_output} from './import-wizard/step_1_2_output';
 
 describe('Forge API tests:', () => {
   let mockLog: any;
@@ -50,7 +48,7 @@ describe('Forge API tests:', () => {
     fabric8ForgeService = TestBed.get(Fabric8ForgeService);
   });
 
-  it('Step_1_1 - init - import wizard: GET command successfully', done => {
+  it('Step_1_1 - import wizard: init', done => {
     // given
     provider.addInteraction({
       state: 'step1.1.init',
@@ -62,7 +60,7 @@ describe('Forge API tests:', () => {
       willRespondWith: {
         status: 200,
         headers: { 'Content-Type': 'application/json' },
-        body: step_1_1_response
+        body: step_1_1_output
       }
     }).then(() => done(), function(err) { console.log('failed'); done.fail(err); });
     const log = () => { };
@@ -83,7 +81,7 @@ describe('Forge API tests:', () => {
     });
   });
 
-  it('Step_1_2 - validate - import wizard: POST command successfully', done => {
+  it('Step_1_2 - import wizard: validate git organisation', done => {
     // given
     provider.addInteraction({
       state: 'step1.2.validate',
@@ -95,7 +93,7 @@ describe('Forge API tests:', () => {
       willRespondWith: {
         status: 200,
         headers: { 'Content-Type': 'application/json' },
-        body: step_1_2_response
+        body: step_1_2_output
       }
     }).then(() => done(), function(err) { console.log('failed'); done.fail(err); });
     const log = () => { };
@@ -117,7 +115,7 @@ describe('Forge API tests:', () => {
     });
   });
 
-  it('Step_1_3 - next - import wizard: POST command successfully', done => {
+  it('Step_1_3 - import wizard: next git organisation', done => {
     // given
     provider.addInteraction({
       state: 'step1.3.next',
@@ -129,7 +127,7 @@ describe('Forge API tests:', () => {
       willRespondWith: {
         status: 200,
         headers: { 'Content-Type': 'application/json' },
-        body: step_1_3_response
+        body: step_1_3_output
       }
     }).then(() => done(), function(err) { console.log('failed'); done.fail(err); });
     const log = () => { };
