+++ src/tests/forge-api/import-wizard/step_1_1_input.ts
@@ -1,0 +1,1 @@
+const step_1_1_input = {};
